refactor : 검색 레포지토리 분리 (#86)
