+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/search/service/SearchService.java
@@ -10,10 +10,10 @@ import com.breaditnow.customer.domain.search.controller.req.SearchRequest;
 import com.breaditnow.customer.domain.search.controller.res.SearchBakeryPageResponse;
 import com.breaditnow.customer.domain.search.controller.res.SearchProductPageResponse;
 import com.breaditnow.customer.domain.search.controller.res.SearchProductResponse;
-import com.breaditnow.domain.domain.bakery.repository.BakeryRepository;
+import com.breaditnow.domain.domain.bakery.repository.SearchBakeryRepository;
 import com.breaditnow.domain.domain.favorite.repository.customerproductfavorite.CustomerProductFavoriteRepository;
 import com.breaditnow.domain.domain.product.entity.Product;
-import com.breaditnow.domain.domain.product.repository.ProductRepository;
+import com.breaditnow.domain.domain.product.repository.SearchProductRepository;
 import com.breaditnow.domain.global.dto.BakeryDistanceDto;
 import com.breaditnow.domain.global.dto.GeoPoint;
 
@@ -23,8 +23,8 @@ import lombok.RequiredArgsConstructor;
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class SearchService {
-	private final BakeryRepository bakeryRepository;
-	private final ProductRepository productRepository;
+	private final SearchBakeryRepository bakeryRepository;
+	private final SearchProductRepository productRepository;
 	private final CustomerProductFavoriteRepository favoriteRepository;
 
 	public SearchBakeryPageResponse searchBakeries(Long customerId, SearchRequest searchRequest) {
@@ -42,8 +42,8 @@ public class SearchService {
 		Pageable pageable = PageRequest.of(searchRequest.page(), searchRequest.size());
 		GeoPoint geoPoint = GeoPoint.of(searchRequest.latitude(), searchRequest.longitude());
 
-		Page<Product> productPage = productRepository.searchProductsWithKeyword(customerId, pageable,
-			searchRequest.sort(), searchRequest.keyword(), geoPoint);
+		Page<Product> productPage = productRepository.searchProductsWithKeyword(pageable, searchRequest.sort(),
+			searchRequest.keyword(), geoPoint);
 
 		Page<SearchProductResponse> searchProductResponses = productPage.map(product -> {
 			boolean isFavorite = favoriteRepository.existsByCustomerIdAndProductId(customerId, product.getId());
