+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/bakery/repository/BakeryRepositoryCustom.java
@@ -3,7 +3,6 @@ package com.breaditnow.domain.domain.bakery.repository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import com.breaditnow.domain.domain.bakery.enumerate.SortType;
 import com.breaditnow.domain.global.dto.BakeryDistanceDto;
 import com.breaditnow.domain.global.dto.GeoPoint;
 
@@ -11,7 +10,4 @@ public interface BakeryRepositoryCustom {
 	Page<BakeryDistanceDto> searchHotBakeriesByFavorite(Long customerId, Pageable pageable, GeoPoint geoPoint);
 
 	Page<BakeryDistanceDto> searchHotBakeriesByReservation(Long cuwstomerId, Pageable pageable, GeoPoint geoPoint);
-
-	Page<BakeryDistanceDto> searchBakeriesWithKeyword(Long customerId, Pageable pageable, SortType sort, String keyword,
-		GeoPoint geoPoint);
 }
