+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/product/repository/ProductRepositoryImpl.java
@@ -1,7 +1,6 @@
 package com.breaditnow.domain.domain.product.repository;
 
 import static com.breaditnow.domain.domain.bakery.entity.QBakery.*;
-import static com.breaditnow.domain.domain.bakery.enumerate.SortType.*;
 import static com.breaditnow.domain.domain.customer.entity.QCustomerRegionPreference.*;
 import static com.breaditnow.domain.domain.favorite.entity.QCustomerProductFavorite.*;
 import static com.breaditnow.domain.domain.product.entity.QProduct.*;
@@ -21,12 +20,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 
-import com.breaditnow.domain.domain.bakery.enumerate.SortType;
 import com.breaditnow.domain.domain.product.entity.Product;
-import com.breaditnow.domain.global.dto.GeoDistanceExpressionProvider;
-import com.breaditnow.domain.global.dto.GeoPoint;
 import com.querydsl.core.types.dsl.BooleanExpression;
-import com.querydsl.core.types.dsl.NumberExpression;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -36,7 +31,6 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ProductRepositoryImpl implements ProductRepositoryCustom {
 	private final JPAQueryFactory queryFactory;
-	private final GeoDistanceExpressionProvider distanceExpressionProvider;
 
 	@Override
 	public Page<Product> searchHotProductsByFavorite(Long customerId, Pageable pageable) {
@@ -100,43 +94,6 @@ public class ProductRepositoryImpl implements ProductRepositoryCustom {
 		List<Product> fetchedList = fetchProductsWithBakery(productIds);
 		List<Product> sortedContent = sortProductsByIds(fetchedList, productIds);
 		return new PageImpl<>(sortedContent, pageable, totalCount);
-	}
-
-	@Override
-	public Page<Product> searchProductsWithKeyword(Long customerId, Pageable pageable, SortType sort, String keyword,
-		GeoPoint geoPoint) {
-
-		BooleanExpression baseCondition = product.isActive.eq(true)
-			.and(product.isHidden.eq(false))
-			.and(product.name.containsIgnoreCase(keyword));
-
-		JPAQuery<Product> query = queryFactory
-			.selectFrom(product)
-			.leftJoin(bakery).on(product.bakery.eq(bakery))
-			.where(baseCondition);
-
-		if (sort == POPULAR) {
-			query.leftJoin(customerProductFavorite).on(customerProductFavorite.product.eq(product))
-				.groupBy(product.id)
-				.orderBy(customerProductFavorite.count().desc(), product.id.asc());
-		} else {
-			if (sort == LATEST) {
-				query.orderBy(product.modifiedAt.desc());
-			} else if (sort == DISTANCE) {
-				NumberExpression<Double> distanceExpression = distanceExpressionProvider.buildDistanceExpression(
-					geoPoint, bakery);
-				query.orderBy(distanceExpression.asc());
-			} else {
-				query.orderBy(product.modifiedAt.desc());
-			}
-		}
-
-		Long totalCount = queryFactory.select(product.count())
-			.from(product)
-			.where(baseCondition)
-			.fetchOne();
-
-		return new PageImpl<>(query.fetch(), pageable, totalCount == null ? 0 : totalCount);
 	}
 
 	private Long fetchTotalCount(Long customerId) {
